signal/requirements.txt
# Core ML
sentence-transformers==2.7.0
bertopic==0.16.3
umap-learn==0.5.6
hdbscan==0.8.40
scikit-learn==1.5.0

# Data
pandas==2.2.2
numpy==1.26.4
pyarrow==16.1.0
duckdb==1.0.0

# NLP / stance
transformers==4.42.0
torch==2.3.1

# Graph
networkx==3.3
python-louvain==0.16

# Embeddings index
faiss-cpu==1.8.0

# Utils
tqdm==4.66.4
python-dotenv==1.0.1
requests==2.32.3